+++ src/pydidas/widgets/selection/__init__.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023 - 2025, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2026, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -20,7 +20,7 @@ Package with widgets which allow the selection of a specific element.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023 - 2025, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2026, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -30,6 +30,7 @@ from .configure_binary_decoding_widget import ConfigureBinaryDecodingWidget
 from .directory_explorer import DirectoryExplorer
 from .hdf5_dataset_selector import Hdf5DatasetSelector
 from .select_data_frame_widget import SelectDataFrameWidget
+from .toggle_options_button import ToggleOptionsButton
 
 
 __all__ = [
@@ -37,4 +38,5 @@ __all__ = [
     "DirectoryExplorer",
     "Hdf5DatasetSelector",
     "SelectDataFrameWidget",
+    "ToggleOptionsButton",
 ]